Continued sequencies.

+++ java/madnet/sequence/IBuffer.java
@@ -2,6 +2,6 @@ package madnet.sequence;
 
 public interface IBuffer
 {
-    ISequence sequence(long offset, long size);
-    long size();
+    ISequence sequence(int offset, int size);
+    int size();
 }
